# docs/swagger.yml
swagger: "2.0"
info:
  title: "REST API Documentatie WP3"
  description: >
    In onze REST API maken we gebruik van JWT voor authenticatie. Alle beschermde routes vereisen een "access_token". Deze wordt gegenereerd bij elke succesvolle inlogpoging.
  version: "2.0"
...
paths:
  /login:
    post:
      tags:
        - "Authentication"
      summary: "Authenticeert de inlogpoging van een ervaringsdeskundige, organisatie of admin"
      description: |
        # Voorbeeld ervaringsdeskundige:
        **email**: demo@ervaringsdeskundige.nl
        **password**: demo123
        **rol**: ervaringsdeskundige

        # Voorbeeld organisatie:
        **email**: demo@organisatie.nl
        **password**: demo123
        **rol**: organisatie

        # Voorbeeld admin:
        **email**: demo@admin.nl
        **password**: demo123
        **rol**: admin
      parameters:
        - name: "body"
          in: "body"
          description: "E-mail, wachtwoord en rol van de gebruiker"
          required: true
          schema:
            type: "object"
            required:
              - email
              - password
              - role
            properties:
              email:
                type: "string"
                example: "demo@ervaringsdeskundige.nl"
                description: "E-mail van gebruiker."
              password:
                type: "string"
                example: "demo123"
                description: "Wachtwoord van gebruiker."
              role:
                type: "string"
                enum: ["ervaringsdeskundige", "organisatie", "admin"]
                example: "ervaringsdeskundige"
                description: "Rol van de gebruiker ('ervaringsdeskundige', 'organisatie', 'admin')."
      responses:
        200:
          description: "Login succesvol"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: true
              message:
                type: "string"
                example: "string"
              redirect:
                type: "string"
                example: "string"
              access_token:
                type: "string"
                example: "string"
        401:
          description: "Ongeldige inloggegevens"
        403:
          description: "Registratie afgekeurd of nog niet goedgekeurd"
  /logout:
    get:
      tags:
        - "Authentication"
      summary: "Verwerkt de uitlogpoging van een ervaringsdeskundige, organisatie of admin"
      description: "Bij het uitloggen sturen we een response terug die de access_token niet langer geldig maakt en een succesvolle uitlogpoging vaststelt."
      responses:
        200:
          description: "Logout succesvol"
          schema:
            $ref: "#/definitions/ApiResponseLink"
        401:
          description: "Geen JWT gevonden in cookies of headers, uitloggen is niet mogelijk zonder eerst ingelogd te zijn"
      security:
        - jwt_auth: []
  /api/users:
    get:
      tags:
        - "Users"
      summary: "Geeft een lijst van alle ervaringsdeskundigen"
      description: "Geeft een lijst van alle ervaringsdeskundigen met betrekking tot de optionele argumenten en rol van degene die de request maakt."
      parameters:
        - name: zoekterm
          in: query
          type: string
          required: false
          description: Pas de weergave van ervaringsdeskundigen aan op basis van deze invoer.
        - name: sorteer_op
          in: query
          type: string
          required: false
          description: Sorteer de weergave van ervaringsdeskundigen op basis van deze invoer.
        - name: volgorde
          in: query
          type: string
          required: false
          description: Op- of aflopend
      responses:
        200:
          description: "Alle ervaringsdeskundigen opgehaald"
          schema:
            type: "object"
            properties:
              current_page:
                type: integer
                example: 1
              data:
                type: "array"
                items:
                  $ref: "#/definitions/GetUser"
              message:
                type: "string"
                example: "string"
              success:
                type: "boolean"
                example: true
              total_pages:
                type: integer
                example: 1
              total_records:
                type: integer
                example: 1
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Geen toegang tot deze informatie"
      security:
        - jwt_auth: []

    post:
      tags:
        - "Users"
      summary: "Verwerkt de registratiepoging van een ervaringsdeskundige"
      description: "Bij een succesvolle registratie worden de gegevens opgeslagen in de database."
      parameters:
        - name: "body"
          in: "body"
          description: |
            **Verplichte invoervelden**
            email, wachtwoord, voornaam, achternaam, geboortedatum, geslacht, postcode, telefoonnummer, introductie, contact_voorkeur, heeft_voogd, voorwaarden_akkoord
            **Optionele invoervelden**
            beperkingen, hulpmiddelen, onderzoektype_voorkeur, (*indien van toepassing*, naam_voogd, email_voogd, telefoonnummer_voogd) 
          required: true
          schema:
            $ref: "#/definitions/CreateUser"
      responses:
        201:
          description: "Registratie van de ervaringsdeskundige succesvol opgeslagen"
          schema:
            $ref: "#/definitions/ApiResponseLink"
        500:
          description: "Registratie mislukt"
  /api/users/{user_id}:
    get:
      tags:
        - "Users"
      summary: "Geeft alle gegevens van de ervaringsdeskundige met de opgegeven user_id"
      description: "Haalt de gegevens op van de ervaringsdeskundige aan de hand van de user_id"
      parameters:
      - name: "user_id"
        in: "path"
        description: "De unieke ID van een ervaringsdeskundige"
        required: true
        type: "string"
      responses:
        200:
          description: "Gebruikergegevens succesvol opgehaald"
          schema:
            type: "object"
            properties:
              data:
                type: "object"
                $ref: "#/definitions/GetUser"
              message:
                type: "string"
                example: "string"
              success:
                type: "boolean"
                example: true
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Geen toegang tot deze informatie"
        404:
          description: "Ervaringsdeskundige niet gevonden"
      security:
        - jwt_auth: []

    put:
      tags:
        - "Users"
      summary: "Past gegevens aan van een ervaringsdeskundige met de opgegeven user_id"
      description: "Past gegevens aan van een ervaringsdeskundige met de opgegeven user_id"
      parameters:
      - name: "user_id"
        in: "path"
        description: "De unieke ID van een ervaringsdeskundige"
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: |
          **Verplichte invoervelden**
          email, wachtwoord, voornaam, achternaam, geboortedatum, geslacht, postcode, telefoonnummer, introductie, contact_voorkeur, heeft_voogd, voorwaarden_akkoord
          **Optionele invoervelden**
          beperkingen, hulpmiddelen, onderzoektype_voorkeur, (*indien van toepassing*, naam_voogd, email_voogd, telefoonnummer_voogd) 
        schema:
          $ref: "#/definitions/CreateUser"
        required: true
      responses:
        200:
          description: "Gebruikergegevens succesvol aangepast"
          schema:
            $ref: "#/definitions/ApiResponseLink"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Niet toegestaan om de gegevens van een andere ervaringsdeskundige aan te passen"
      security:
        - jwt_auth: []

    patch:
      tags:
        - "Users"
      summary: "Gebruik om alleen de status van de ervaringsdeskundige met de opgegeven user_id aan te passen"
      description: "Gebruik om alleen de status van de ervaringsdeskundige met de opgegeven user_id aan te passen"
      parameters:
      - name: "user_id"
        in: "path"
        description: De unieke ID van een ervaringsdeskundige
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: "Alleen de status van een ervaringsdeskundige"
        required: true
        schema:
          type: "object"
          required:
              - status
          properties:
            status:
              type: "string"
              example: "string"
              enum: ["Nieuw", "Goedgekeurd", "Afgekeurd"]
              description: "Status van een ervaringsdeskundige"
      responses:
        200:
          description: "Wijzigingen zijn succesvol opgeslagen"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Niet toegestaan om de gegevens van een andere ervaringsdeskundige aan te passen"
      security:
        - jwt_auth: []

    delete:
      tags:
        - "Users"
      summary: "Verwijdert alle gegevens van een ervaringsdeskundige met de opgegeven user_id"
      description: "Verwijdert alle gegevens van een ervaringsdeskundige met de opgegeven user_id"
      parameters:
      - name: "user_id"
        in: "path"
        description: "De unieke ID van een ervaringsdeskundige"
        required: true
        type: "string"
      responses:
        200:
          description: "Ervaringsdeskundige succesvol verwijderd"
          schema:
            $ref: "#/definitions/ApiResponseLink"
        401:
          description: "Geen JWT gevonden in cookies of headers, zonder authorisatie geen toegang"
        403:
          description: "Niet toegestaan om de gegevens van een andere ervaringsdeskundige te verwijderen"
      security:
        - jwt_auth: []
...
securityDefinitions:
  jwt_auth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Voer in: Bearer <access_token>

      Voorbeeld: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
definitions:
  ApiResponse:
    type: "object"
    properties:
      success:
        type: "boolean"
        example: true
      message:
        type: "string"
        example: "string"
  ApiResponseLink:
    type: "object"
    properties:
      success:
        type: "boolean"
        example: true
      message:
        type: "string"
        example: "string"
      redirect:
        type: "string"
        example: "string"
...
  CreateUser:
    type: "object"
    properties:
      achternaam:
        type: "string"
        example: "string"
      beperkingen:
        type: "array"
        items:
          type: integer
      bijzonderheden_beperkingen:
        type: "string"
        example: "string"
      bijzonderheden_beschikbaarheid:
        type: "string"
        example: "string"
      contact_voorkeur:
        type: "string"
        example: "string"
      email:
        type: "string"
        example: "string"
      email_voogd:
        type: "string"
        example: "string"
      geboortedatum:
        type: "string"
        example: "string"
      geslacht:
        type: "string"
        example: "string"
      heeft_voogd:
        type: "integer"
        example: "0"
      hulpmiddelen:
        type: "string"
        example: "string"
      introductie:
        type: "string"
        example: "string"
      name_voogd:
        type: "string"
        example: "string"
      onderzoektype_voorkeur:
        type: "string"
        example: "string"
      postcode:
        type: "string"
        example: "string"
      telefoonnummer:
        type: "string"
        example: "string"
      telefoonnummer_voogd:
        type: "string"
        example: "string"
      tussenvoegsel:
        type: "string"
        example: "string"
      voornaam:
        type: "string"
        example: "string"
      voorwaarden_akkoord:
        type: "integer"
        example: "0"
      wachtwoord:
        type: "string"
        example: "string"
    xml:
      name: "CreateUser"
  GetUser:
    type: "object"
    properties:
      achternaam:
        type: "string"
        example: "string"
      beheerder:
        type: "integer"
        example: "0"
      bijzonderheden_beperkingen:
        type: "string"
        example: "string"
      bijzonderheden_beschikbaarheid:
        type: "string"
        example: "string"
      contact_voorkeur:
        type: "string"
        example: "string"
      email:
        type: "string"
        example: "string"
      email_voogd:
        type: "string"
        example: "string"
      geboortedatum:
        type: "string"
        example: "string"
      gebruiker_id:
        type: "integer"
        example: "0"
      geslacht:
        type: "string"
        example: "string"
      heeft_voogd:
        type: "integer"
        example: "0"
      hulpmiddelen:
        type: "string"
        example: "string"
      introductie:
        type: "string"
        example: "string"
      name_voogd:
        type: "string"
        example: "string"
      onderzoektype_voorkeur:
        type: "string"
        example: "string"
      postcode:
        type: "string"
        example: "string"
      status:
        type: "string"
        example: "string"
      telefoonnummer:
        type: "string"
        example: "string"
      telefoonnummer_voogd:
        type: "string"
        example: "string"
      tussenvoegsel:
        type: "string"
        example: "string"
      voornaam:
        type: "string"
        example: "string"
      voorwaarden_akkoord:
        type: "integer"
        example: "0"
      wachtwoord:
        type: "string"
        example: "string"
    xml:
      name: "GetUser"
...